school_backend/requirements.txt
# Dependencias del Sistema Escolar Backend
# FastAPI y servidor
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Base de datos
SQLAlchemy>=2.0.0
PyMySQL>=1.1.0

# Autenticación y seguridad
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.1
cryptography>=41.0.0

# Validación de datos
pydantic>=2.0.0
pydantic-settings>=2.0.0
email-validator>=2.0.0

# Utilidades
python-multipart>=0.0.6

# Testing (opcional)
pytest>=7.4.0
httpx>=0.25.0
pytest-asyncio>=0.21.0

